requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.9
pillow==10.4.0
pytesseract==0.3.13
pdf2image==1.17.0
httpx==0.27.2
python-magic==0.4.27
google-generativeai==0.7.2
python-dotenv==1.0.1

